Give more details if pipeline can't be set to PAUSED
When the pipeline can't be set to PAUSED, it can be helpful to provide
the user with more details. Example attempting to process a non-existent
file:

```
Error: Failed setting pipeline to PAUSED
        Resource not found.
        ../subprojects/gstreamer/plugins/elements/gstfilesrc.c(585): gst_file_src_start (): /GstPipeline:pipeline0/GstURIDecodeBin:uridecodebin0/GstFileSrc:source:
No such file "/home/fengalin/temp/Doesn't exist.mp4"
```

Of course the final message could be parsed / improved further, but
that's a good start at least.