Set number of threads to 1 for common decoders
Ported from Robert Mader's Totem patch:
https://gitlab.gnome.org/rmader/totem/-/commit/38ae0f8158cd13f7306b32cd3db73dc8b225f2d1

Various decoders default to the number of CPU cores (or threads) for
decoder threads, requiring the allocation of output buffers for each
of them. On system with many cores this can cause the allocations to
exceed the RLIMIT_DATA memory limit.

Set the number of threads to 1. That should be all we need for the
thumbnailer use-case, preventing the mentioned cases and generally
reducing memory usage/allocations.

Unfortunately there isn't a generic way to do so, thus use
element-setup callback for well known codecs/properties for now.

See also
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/4423